Move sub-classes of "HtmlDocument", "HtmlElement" and "HtmlWindow" classes to their own files
