+++ src/System.Windows.Forms/src/System/Windows/Forms/HtmlDocument.cs
@@ -14,7 +14,7 @@ using static Interop.Mshtml;
 
 namespace System.Windows.Forms
 {
-    public sealed class HtmlDocument
+    public sealed partial class HtmlDocument
     {
         internal static object EventClick = new object();
         internal static object EventContextMenuShowing = new object();
@@ -599,316 +599,6 @@ namespace System.Windows.Forms
             }
 
             return Color.Empty;
-        }
-
-        /// <summary>
-        ///  HtmlDocumentShim - this is the glue between the DOM eventing mechanisms
-        ///          and our CLR callbacks.
-        ///
-        ///  There are two kinds of events: HTMLWindowEvents2 and IHtmlWindow3.AttachHandler style
-        ///  HTMLDocumentEvents2: we create an IConnectionPoint (via ConnectionPointCookie) between us and MSHTML and it calls back
-        ///              on our an instance of HTMLDocumentEvents2.  The HTMLDocumentEvents2 class then fires the event.
-        ///
-        ///  IHTMLDocument3.AttachHandler: MSHML calls back on an HtmlToClrEventProxy that we've created, looking
-        ///                       for a method named DISPID=0.  For each event that's subscribed, we create
-        ///                       a new HtmlToClrEventProxy, detect the callback and fire the corresponding
-        ///                       CLR event.
-        /// </summary>
-        internal class HtmlDocumentShim : HtmlShim
-        {
-            private AxHost.ConnectionPointCookie cookie;
-            private HtmlDocument htmlDocument;
-            private readonly IHTMLWindow2 associatedWindow;
-
-            internal HtmlDocumentShim(HtmlDocument htmlDocument)
-            {
-                this.htmlDocument = htmlDocument;
-                // snap our associated window so we know when to disconnect.
-                if (this.htmlDocument != null)
-                {
-                    HtmlWindow window = htmlDocument.Window;
-                    if (window != null)
-                    {
-                        associatedWindow = window.NativeHtmlWindow;
-                    }
-                }
-            }
-
-            public override IHTMLWindow2 AssociatedWindow
-            {
-                get { return associatedWindow; }
-            }
-
-            public IHTMLDocument2 NativeHtmlDocument2
-            {
-                get { return htmlDocument.NativeHtmlDocument2; }
-            }
-
-            internal HtmlDocument Document
-            {
-                get { return htmlDocument; }
-            }
-
-            ///  Support IHtmlDocument3.AttachHandler
-            public override void AttachEventHandler(string eventName, EventHandler eventHandler)
-            {
-                // IE likes to call back on an IDispatch of DISPID=0 when it has an event,
-                // the HtmlToClrEventProxy helps us fake out the CLR so that we can call back on
-                // our EventHandler properly.
-
-                HtmlToClrEventProxy proxy = AddEventProxy(eventName, eventHandler);
-                ((IHTMLDocument3)NativeHtmlDocument2).AttachEvent(eventName, proxy);
-            }
-
-            ///  Support IHtmlDocument3.DetachHandler
-            public override void DetachEventHandler(string eventName, EventHandler eventHandler)
-            {
-                HtmlToClrEventProxy proxy = RemoveEventProxy(eventHandler);
-                if (proxy != null)
-                {
-                    ((IHTMLDocument3)NativeHtmlDocument2).DetachEvent(eventName, proxy);
-                }
-            }
-
-            //
-            // Connect to standard events
-            //
-            public override void ConnectToEvents()
-            {
-                if (cookie is null || !cookie.Connected)
-                {
-                    cookie = new AxHost.ConnectionPointCookie(NativeHtmlDocument2,
-                                                                          new HTMLDocumentEvents2(htmlDocument),
-                                                                          typeof(DHTMLDocumentEvents2),
-                                                                          /*throwException*/ false);
-                    if (!cookie.Connected)
-                    {
-                        cookie = null;
-                    }
-                }
-            }
-
-            //
-            // Disconnect from standard events
-            //
-            public override void DisconnectFromEvents()
-            {
-                if (cookie != null)
-                {
-                    cookie.Disconnect();
-                    cookie = null;
-                }
-            }
-
-            protected override void Dispose(bool disposing)
-            {
-                base.Dispose(disposing);
-                if (disposing)
-                {
-                    if (htmlDocument != null)
-                    {
-                        Marshal.FinalReleaseComObject(htmlDocument.NativeHtmlDocument2);
-                    }
-
-                    htmlDocument = null;
-                }
-            }
-
-            protected override object GetEventSender()
-            {
-                return htmlDocument;
-            }
-        }
-
-        //
-        // Private classes:
-        //
-        [ClassInterface(ClassInterfaceType.None)]
-        private class HTMLDocumentEvents2 : StandardOleMarshalObject, /*Enforce calling back on the same thread*/
-                                            DHTMLDocumentEvents2
-        {
-            private readonly HtmlDocument parent;
-
-            public HTMLDocumentEvents2(HtmlDocument htmlDocument)
-            {
-                parent = htmlDocument;
-            }
-
-            private void FireEvent(object key, EventArgs e)
-            {
-                if (parent != null)
-                {
-                    parent.DocumentShim.FireEvent(key, e);
-                }
-            }
-
-            public bool onclick(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventClick, e);
-
-                return e.ReturnValue;
-            }
-
-            public bool oncontextmenu(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventContextMenuShowing, e);
-                return e.ReturnValue;
-            }
-
-            public void onfocusin(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventFocusing, e);
-            }
-
-            public void onfocusout(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventLosingFocus, e);
-            }
-
-            public void onmousemove(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventMouseMove, e);
-            }
-
-            public void onmousedown(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventMouseDown, e);
-            }
-
-            public void onmouseout(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventMouseLeave, e);
-            }
-
-            public void onmouseover(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventMouseOver, e);
-            }
-
-            public void onmouseup(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventMouseUp, e);
-            }
-
-            public bool onstop(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlDocument.EventStop, e);
-                return e.ReturnValue;
-            }
-
-            public bool onhelp(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool ondblclick(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onkeydown(IHTMLEventObj evtObj) { }
-
-            public void onkeyup(IHTMLEventObj evtObj) { }
-
-            public bool onkeypress(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onreadystatechange(IHTMLEventObj evtObj) { }
-
-            public bool onbeforeupdate(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onafterupdate(IHTMLEventObj evtObj) { }
-
-            public bool onrowexit(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onrowenter(IHTMLEventObj evtObj) { }
-
-            public bool ondragstart(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onselectstart(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onerrorupdate(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onrowsdelete(IHTMLEventObj evtObj) { }
-
-            public void onrowsinserted(IHTMLEventObj evtObj) { }
-
-            public void oncellchange(IHTMLEventObj evtObj) { }
-
-            public void onpropertychange(IHTMLEventObj evtObj) { }
-
-            public void ondatasetchanged(IHTMLEventObj evtObj) { }
-
-            public void ondataavailable(IHTMLEventObj evtObj) { }
-
-            public void ondatasetcomplete(IHTMLEventObj evtObj) { }
-
-            public void onbeforeeditfocus(IHTMLEventObj evtObj) { }
-
-            public void onselectionchange(IHTMLEventObj evtObj) { }
-
-            public bool oncontrolselect(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onmousewheel(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onactivate(IHTMLEventObj evtObj) { }
-
-            public void ondeactivate(IHTMLEventObj evtObj) { }
-
-            public bool onbeforeactivate(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onbeforedeactivate(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
         }
 
         #region operators
