+++ src/System.Windows.Forms/src/System/Windows/Forms/HtmlElement.cs
@@ -12,7 +12,7 @@ using static Interop.Mshtml;
 
 namespace System.Windows.Forms
 {
-    public sealed class HtmlElement
+    public sealed partial class HtmlElement
     {
         internal static readonly object EventClick = new object();
         internal static readonly object EventDoubleClick = new object();
@@ -651,510 +651,6 @@ namespace System.Windows.Forms
         {
             add => ElementShim.AddHandler(EventMouseLeave, value);
             remove => ElementShim.RemoveHandler(EventMouseLeave, value);
-        }
-
-        //
-        // Private classes:
-        //
-        [ClassInterface(ClassInterfaceType.None)]
-        private class HTMLElementEvents2 : StandardOleMarshalObject, /*Enforce calling back on the same thread*/
-                                           DHTMLElementEvents2,
-                                           DHTMLAnchorEvents2,
-                                           DHTMLAreaEvents2,
-                                           DHTMLButtonElementEvents2,
-                                           DHTMLControlElementEvents2,
-                                           DHTMLFormElementEvents2,
-                                           DHTMLFrameSiteEvents2,
-                                           DHTMLImgEvents2,
-                                           DHTMLInputFileElementEvents2,
-                                           DHTMLInputImageEvents2,
-                                           DHTMLInputTextElementEvents2,
-                                           DHTMLLabelEvents2,
-                                           DHTMLLinkElementEvents2,
-                                           DHTMLMapEvents2,
-                                           DHTMLMarqueeElementEvents2,
-                                           DHTMLOptionButtonElementEvents2,
-                                           DHTMLSelectElementEvents2,
-                                           DHTMLStyleElementEvents2,
-                                           DHTMLTableEvents2,
-                                           DHTMLTextContainerEvents2,
-                                           DHTMLScriptEvents2
-        {
-            private readonly HtmlElement parent;
-
-            public HTMLElementEvents2(HtmlElement htmlElement)
-            {
-                parent = htmlElement;
-            }
-
-            private void FireEvent(object key, EventArgs e)
-            {
-                if (parent != null)
-                {
-                    parent.ElementShim.FireEvent(key, e);
-                }
-            }
-
-            public bool onclick(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventClick, e);
-                return e.ReturnValue;
-            }
-
-            public bool ondblclick(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventDoubleClick, e);
-                return e.ReturnValue;
-            }
-
-            public bool onkeypress(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventKeyPress, e);
-                return e.ReturnValue;
-            }
-
-            public void onkeydown(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventKeyDown, e);
-            }
-
-            public void onkeyup(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventKeyUp, e);
-            }
-
-            public void onmouseover(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventMouseOver, e);
-            }
-
-            public void onmousemove(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventMouseMove, e);
-            }
-
-            public void onmousedown(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventMouseDown, e);
-            }
-
-            public void onmouseup(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventMouseUp, e);
-            }
-
-            public void onmouseenter(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventMouseEnter, e);
-            }
-
-            public void onmouseleave(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventMouseLeave, e);
-            }
-
-            public bool onerrorupdate(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onfocus(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventGotFocus, e);
-            }
-
-            public bool ondrag(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventDrag, e);
-                return e.ReturnValue;
-            }
-
-            public void ondragend(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventDragEnd, e);
-            }
-
-            public void ondragleave(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventDragLeave, e);
-            }
-
-            public bool ondragover(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventDragOver, e);
-                return e.ReturnValue;
-            }
-
-            public void onfocusin(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventFocusing, e);
-            }
-
-            public void onfocusout(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventLosingFocus, e);
-            }
-
-            public void onblur(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlElement.EventLostFocus, e);
-            }
-
-            public void onresizeend(IHTMLEventObj evtObj) { }
-
-            public bool onresizestart(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onhelp(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onmouseout(IHTMLEventObj evtObj) { }
-
-            public bool onselectstart(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onfilterchange(IHTMLEventObj evtObj) { }
-
-            public bool ondragstart(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onbeforeupdate(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onafterupdate(IHTMLEventObj evtObj) { }
-
-            public bool onrowexit(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onrowenter(IHTMLEventObj evtObj) { }
-
-            public void ondatasetchanged(IHTMLEventObj evtObj) { }
-
-            public void ondataavailable(IHTMLEventObj evtObj) { }
-
-            public void ondatasetcomplete(IHTMLEventObj evtObj) { }
-
-            public void onlosecapture(IHTMLEventObj evtObj) { }
-
-            public void onpropertychange(IHTMLEventObj evtObj) { }
-
-            public void onscroll(IHTMLEventObj evtObj) { }
-
-            public void onresize(IHTMLEventObj evtObj) { }
-
-            public bool ondragenter(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool ondrop(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onbeforecut(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool oncut(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onbeforecopy(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool oncopy(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onbeforepaste(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onpaste(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool oncontextmenu(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onrowsdelete(IHTMLEventObj evtObj) { }
-
-            public void onrowsinserted(IHTMLEventObj evtObj) { }
-
-            public void oncellchange(IHTMLEventObj evtObj) { }
-
-            public void onreadystatechange(IHTMLEventObj evtObj) { }
-
-            public void onlayoutcomplete(IHTMLEventObj evtObj) { }
-
-            public void onpage(IHTMLEventObj evtObj) { }
-
-            public void onactivate(IHTMLEventObj evtObj) { }
-
-            public void ondeactivate(IHTMLEventObj evtObj) { }
-
-            public bool onbeforedeactivate(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onbeforeactivate(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onmove(IHTMLEventObj evtObj) { }
-
-            public bool oncontrolselect(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onmovestart(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onmoveend(IHTMLEventObj evtObj) { }
-
-            public bool onmousewheel(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onchange(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onselect(IHTMLEventObj evtObj) { }
-
-            public void onload(IHTMLEventObj evtObj) { }
-
-            public void onerror(IHTMLEventObj evtObj) { }
-
-            public void onabort(IHTMLEventObj evtObj) { }
-
-            public bool onsubmit(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public bool onreset(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onchange_void(IHTMLEventObj evtObj) { }
-
-            public void onbounce(IHTMLEventObj evtObj) { }
-
-            public void onfinish(IHTMLEventObj evtObj) { }
-
-            public void onstart(IHTMLEventObj evtObj) { }
-        }
-
-        /// <summary>
-        ///  HtmlElementShim - this is the glue between the DOM eventing mechanisms
-        ///          and our CLR callbacks.
-        ///
-        ///  HTMLElementEvents2: we create an IConnectionPoint (via ConnectionPointCookie) between us and MSHTML and it calls back
-        ///              on our an instance of HTMLElementEvents2.  The HTMLElementEvents2 class then fires the event.
-        ///
-        /// </summary>
-        internal class HtmlElementShim : HtmlShim
-        {
-            private static readonly Type[] dispInterfaceTypes =
-            {
-                typeof(DHTMLElementEvents2),
-                typeof(DHTMLAnchorEvents2),
-                typeof(DHTMLAreaEvents2),
-                typeof(DHTMLButtonElementEvents2),
-                typeof(DHTMLControlElementEvents2),
-                typeof(DHTMLFormElementEvents2),
-                typeof(DHTMLFrameSiteEvents2),
-                typeof(DHTMLImgEvents2),
-                typeof(DHTMLInputFileElementEvents2),
-                typeof(DHTMLInputImageEvents2),
-                typeof(DHTMLInputTextElementEvents2),
-                typeof(DHTMLLabelEvents2),
-                typeof(DHTMLLinkElementEvents2),
-                typeof(DHTMLMapEvents2),
-                typeof(DHTMLMarqueeElementEvents2),
-                typeof(DHTMLOptionButtonElementEvents2),
-                typeof(DHTMLSelectElementEvents2),
-                typeof(DHTMLStyleElementEvents2),
-                typeof(DHTMLTableEvents2),
-                typeof(DHTMLTextContainerEvents2),
-                typeof(DHTMLScriptEvents2)
-            };
-
-            private AxHost.ConnectionPointCookie cookie;   // To hook up events from the native HtmlElement
-            private HtmlElement htmlElement;
-            private readonly IHTMLWindow2 associatedWindow;
-
-            public HtmlElementShim(HtmlElement element)
-            {
-                htmlElement = element;
-
-                // snap our associated window so we know when to disconnect.
-                if (htmlElement != null)
-                {
-                    HtmlDocument doc = htmlElement.Document;
-                    if (doc != null)
-                    {
-                        HtmlWindow window = doc.Window;
-                        if (window != null)
-                        {
-                            associatedWindow = window.NativeHtmlWindow;
-                        }
-                    }
-                }
-            }
-
-            public IHTMLElement NativeHtmlElement
-            {
-                get { return htmlElement.NativeHtmlElement; }
-            }
-
-            internal HtmlElement Element
-            {
-                get { return htmlElement; }
-            }
-
-            public override IHTMLWindow2 AssociatedWindow
-            {
-                get { return associatedWindow; }
-            }
-
-            ///  Support IHTMLElement2.AttachEventHandler
-            public override void AttachEventHandler(string eventName, EventHandler eventHandler)
-            {
-                // IE likes to call back on an IDispatch of DISPID=0 when it has an event,
-                // the HtmlToClrEventProxy helps us fake out the CLR so that we can call back on
-                // our EventHandler properly.
-
-                HtmlToClrEventProxy proxy = AddEventProxy(eventName, eventHandler);
-                ((IHTMLElement2)NativeHtmlElement).AttachEvent(eventName, proxy);
-            }
-
-            public override void ConnectToEvents()
-            {
-                if (cookie is null || !cookie.Connected)
-                {
-                    for (int i = 0; i < dispInterfaceTypes.Length && cookie is null; i++)
-                    {
-                        cookie = new AxHost.ConnectionPointCookie(NativeHtmlElement,
-                                                                                  new HTMLElementEvents2(htmlElement),
-                                                                                  dispInterfaceTypes[i],
-                                                                                  /*throwException*/ false);
-                        if (!cookie.Connected)
-                        {
-                            cookie = null;
-                        }
-                    }
-                }
-            }
-
-            ///  Support IHTMLElement2.DetachHandler
-            public override void DetachEventHandler(string eventName, EventHandler eventHandler)
-            {
-                HtmlToClrEventProxy proxy = RemoveEventProxy(eventHandler);
-                if (proxy != null)
-                {
-                    ((IHTMLElement2)NativeHtmlElement).DetachEvent(eventName, proxy);
-                }
-            }
-
-            public override void DisconnectFromEvents()
-            {
-                if (cookie != null)
-                {
-                    cookie.Disconnect();
-                    cookie = null;
-                }
-            }
-
-            protected override void Dispose(bool disposing)
-            {
-                base.Dispose(disposing);
-                if (htmlElement != null)
-                {
-                    Marshal.FinalReleaseComObject(htmlElement.NativeHtmlElement);
-                }
-
-                htmlElement = null;
-            }
-
-            protected override object GetEventSender()
-            {
-                return htmlElement;
-            }
         }
 
         #region operators
