+++ src/System.Windows.Forms/src/System/Windows/Forms/HtmlWindow.cs
@@ -11,7 +11,7 @@ using static Interop.Mshtml;
 
 namespace System.Windows.Forms
 {
-    public sealed class HtmlWindow
+    public sealed partial class HtmlWindow
     {
         internal static readonly object EventError = new object();
         internal static readonly object EventGotFocus = new object();
@@ -370,196 +370,6 @@ namespace System.Windows.Forms
         {
             add => WindowShim.AddHandler(EventUnload, value);
             remove => WindowShim.RemoveHandler(EventUnload, value);
-        }
-
-        //
-        // Private classes:
-        //
-        [ClassInterface(ClassInterfaceType.None)]
-        private class HTMLWindowEvents2 : StandardOleMarshalObject, /*Enforce calling back on the same thread*/
-            DHTMLWindowEvents2
-        {
-            private readonly HtmlWindow parent;
-
-            public HTMLWindowEvents2(HtmlWindow htmlWindow)
-            {
-                parent = htmlWindow;
-            }
-
-            private void FireEvent(object key, EventArgs e)
-            {
-                if (parent != null)
-                {
-                    parent.WindowShim.FireEvent(key, e);
-                }
-            }
-
-            public void onfocus(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlWindow.EventGotFocus, e);
-            }
-
-            public void onblur(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlWindow.EventLostFocus, e);
-            }
-
-            public bool onerror(string description, string urlString, int line)
-            {
-                HtmlElementErrorEventArgs e = new HtmlElementErrorEventArgs(description, urlString, line);
-                FireEvent(HtmlWindow.EventError, e);
-                return e.Handled;
-            }
-
-            public void onload(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlWindow.EventLoad, e);
-            }
-
-            public void onunload(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlWindow.EventUnload, e);
-                if (parent != null)
-                {
-                    parent.WindowShim.OnWindowUnload();
-                }
-            }
-
-            public void onscroll(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlWindow.EventScroll, e);
-            }
-
-            public void onresize(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                FireEvent(HtmlWindow.EventResize, e);
-            }
-
-            public bool onhelp(IHTMLEventObj evtObj)
-            {
-                HtmlElementEventArgs e = new HtmlElementEventArgs(parent.ShimManager, evtObj);
-                return e.ReturnValue;
-            }
-
-            public void onbeforeunload(IHTMLEventObj evtObj) { }
-
-            public void onbeforeprint(IHTMLEventObj evtObj) { }
-
-            public void onafterprint(IHTMLEventObj evtObj) { }
-        }
-
-        /// <summary>
-        ///  HtmlWindowShim - this is the glue between the DOM eventing mechanisms
-        ///        and our CLR callbacks.
-        ///
-        ///  There are two kinds of events: HTMLWindowEvents2 and IHtmlWindow3.AttachHandler style
-        ///  HTMLWindowEvents2: we create an IConnectionPoint (via ConnectionPointCookie) between us and MSHTML and it calls back
-        ///              on an instance of HTMLWindowEvents2.  The HTMLWindowEvents2 class then fires the event.
-        ///
-        ///  IHTMLWindow3.AttachHandler: MSHML calls back on an HtmlToClrEventProxy that we've created, looking
-        ///                       for a method named DISPID=0.  For each event that's subscribed, we create
-        ///                       a new HtmlToClrEventProxy, detect the callback and fire the corresponding
-        ///                       CLR event.
-        /// </summary>
-        internal class HtmlWindowShim : HtmlShim
-        {
-            private AxHost.ConnectionPointCookie cookie;
-            private HtmlWindow htmlWindow;
-
-            public HtmlWindowShim(HtmlWindow window)
-            {
-                htmlWindow = window;
-            }
-
-            public IHTMLWindow2 NativeHtmlWindow
-            {
-                get { return htmlWindow.NativeHtmlWindow; }
-            }
-
-            public override IHTMLWindow2 AssociatedWindow
-            {
-                get { return htmlWindow.NativeHtmlWindow; }
-            }
-
-            ///  Support IHtmlDocument3.AttachHandler
-            public override void AttachEventHandler(string eventName, EventHandler eventHandler)
-            {
-                // IE likes to call back on an IDispatch of DISPID=0 when it has an event,
-                // the HtmlToClrEventProxy helps us fake out the CLR so that we can call back on
-                // our EventHandler properly.
-
-                HtmlToClrEventProxy proxy = AddEventProxy(eventName, eventHandler);
-                bool success = ((IHTMLWindow3)NativeHtmlWindow).AttachEvent(eventName, proxy);
-                Debug.Assert(success, "failed to add event");
-            }
-
-            ///  Support HTMLWindowEvents2
-            public override void ConnectToEvents()
-            {
-                if (cookie is null || !cookie.Connected)
-                {
-                    cookie = new AxHost.ConnectionPointCookie(NativeHtmlWindow,
-                                                                              new HTMLWindowEvents2(htmlWindow),
-                                                                              typeof(DHTMLWindowEvents2),
-                                                                              /*throwException*/ false);
-                    if (!cookie.Connected)
-                    {
-                        cookie = null;
-                    }
-                }
-            }
-
-            ///  Support IHTMLWindow3.DetachHandler
-            public override void DetachEventHandler(string eventName, EventHandler eventHandler)
-            {
-                HtmlToClrEventProxy proxy = RemoveEventProxy(eventHandler);
-                if (proxy != null)
-                {
-                    ((IHTMLWindow3)NativeHtmlWindow).DetachEvent(eventName, proxy);
-                }
-            }
-
-            public override void DisconnectFromEvents()
-            {
-                if (cookie != null)
-                {
-                    cookie.Disconnect();
-                    cookie = null;
-                }
-            }
-
-            protected override void Dispose(bool disposing)
-            {
-                base.Dispose(disposing);
-                if (disposing)
-                {
-                    if (htmlWindow != null && htmlWindow.NativeHtmlWindow != null)
-                    {
-                        Marshal.FinalReleaseComObject(htmlWindow.NativeHtmlWindow);
-                    }
-
-                    htmlWindow = null;
-                }
-            }
-
-            protected override object GetEventSender()
-            {
-                return htmlWindow;
-            }
-
-            public void OnWindowUnload()
-            {
-                if (htmlWindow != null)
-                {
-                    htmlWindow.ShimManager.OnWindowUnloaded(htmlWindow);
-                }
-            }
         }
 
         #region operators
